#1050 Separate out cli from run command
