+++ tests/neuro_san_studio/commands/test_run.py
@@ -17,7 +17,6 @@
 """Tests for NeuroSanRunner."""
 
 import os
-import sys
 from collections.abc import Callable
 from collections.abc import Iterable
 from pathlib import Path
@@ -27,10 +26,8 @@ import pytest
 from pytest import CaptureFixture
 from pytest import MonkeyPatch
 
-from neuro_san_studio.commands import init as init_module
 from neuro_san_studio.commands import run as run_module
 from neuro_san_studio.commands.run import NeuroSanRunner
-from neuro_san_studio.commands.run import main
 
 
 class TestNeuroSanRunner:
@@ -206,76 +203,3 @@ class TestNeuroSanRunner:
         monkeypatch.setattr(run_module, "timedinput", _capturing_input)
         self._make_runner()._validate_yes_no_input("Kill processes? ")
         assert seen_prompts == ["Kill processes? "]
-
-
-class TestMainEntryPoint:
-    """Tests for the `main()` console script entry point."""
-
-    @staticmethod
-    def _install_fake_runner(monkeypatch: MonkeyPatch) -> list[str]:
-        """Replace NeuroSanRunner with a recording stand-in and return the call log."""
-        call_order: list[str] = []
-
-        class FakeRunner:  # pylint: disable=too-few-public-methods
-            """Stand-in for NeuroSanRunner that records method calls."""
-
-            def __init__(self) -> None:
-                call_order.append("init")
-
-            def run(self) -> None:
-                """Record that run() was invoked."""
-                call_order.append("run")
-
-        monkeypatch.setattr(run_module, "NeuroSanRunner", FakeRunner)
-        return call_order
-
-    def test_main_with_no_args_shows_help(self, monkeypatch: MonkeyPatch) -> None:
-        """Bare `neuro-san-studio` should show help and exit cleanly without starting the server."""
-        call_order = self._install_fake_runner(monkeypatch)
-        monkeypatch.setattr(sys, "argv", ["neuro-san-studio"])
-        # Typer exits with code 0 after printing help; main() swallows that for clean exits.
-        main()
-        assert not call_order
-
-    def test_main_with_run_subcommand_runs_server(self, monkeypatch: MonkeyPatch) -> None:
-        """Explicit `neuro-san-studio run` should start the server."""
-        call_order = self._install_fake_runner(monkeypatch)
-        monkeypatch.setattr(sys, "argv", ["neuro-san-studio", "run"])
-        main()
-        assert call_order == ["init", "run"]
-
-    def test_main_with_init_subcommand_invokes_init(self, monkeypatch: MonkeyPatch) -> None:
-        """`neuro-san-studio init` should invoke InitCommand and NOT NeuroSanRunner."""
-        runner_call_order = self._install_fake_runner(monkeypatch)
-        init_calls: list[tuple[str | None]] = []
-
-        class FakeInit:  # pylint: disable=too-few-public-methods
-            """Stand-in for InitCommand that records the providers_arg it received."""
-
-            def __init__(self, providers_arg: str | None = None) -> None:
-                init_calls.append((providers_arg,))
-
-            def run(self) -> None:
-                """Record that init.run() was invoked."""
-                init_calls.append(("run",))
-
-        monkeypatch.setattr(init_module, "InitCommand", FakeInit)
-        monkeypatch.setattr(sys, "argv", ["neuro-san-studio", "init", "--providers", "openai,anthropic"])
-        main()
-        assert not runner_call_order
-        assert init_calls == [("openai,anthropic",), ("run",)]
-
-    def test_main_propagates_runner_exceptions(self, monkeypatch: MonkeyPatch) -> None:
-        """Exceptions from NeuroSanRunner().run() should bubble up to the caller."""
-
-        class ExplodingRunner:  # pylint: disable=too-few-public-methods
-            """Runner whose run() raises, to verify main() does not swallow errors."""
-
-            def run(self) -> None:
-                """Raise to simulate a runtime failure."""
-                raise RuntimeError("boom")
-
-        monkeypatch.setattr(run_module, "NeuroSanRunner", ExplodingRunner)
-        monkeypatch.setattr(sys, "argv", ["neuro-san-studio", "run"])
-        with pytest.raises(RuntimeError, match="boom"):
-            main()
